Add more efficient implementation for placeholder sanitizer for adventure
