+++ honey-common/src/dev/shiza/honey/adventure/message/formatter/AdventureMessageFormatterFactory.java
@@ -1,6 +1,7 @@
 package dev.shiza.honey.adventure.message.formatter;
 
 import dev.shiza.honey.adventure.message.compiler.AdventureMessageCompilerFactory;
+import dev.shiza.honey.adventure.placeholder.sanitizer.AdventurePlaceholderSanitizerFactory;
 import dev.shiza.honey.conversion.ImplicitConversion;
 import dev.shiza.honey.message.compiler.MessageCompiler;
 import dev.shiza.honey.placeholder.PlaceholderContext;
@@ -10,7 +11,6 @@ import dev.shiza.honey.placeholder.processor.PlaceholderProcessor;
 import dev.shiza.honey.placeholder.processor.PlaceholderProcessorFactory;
 import dev.shiza.honey.placeholder.resolver.PlaceholderResolver;
 import dev.shiza.honey.placeholder.resolver.PlaceholderResolverFactory;
-import dev.shiza.honey.placeholder.sanitizer.DelegatingPlaceholderSanitizerFactory;
 import dev.shiza.honey.placeholder.sanitizer.PlaceholderSanitizer;
 import dev.shiza.honey.processor.ProcessorRegistry;
 import dev.shiza.honey.processor.ProcessorRegistryFactory;
@@ -62,7 +62,7 @@ public final class AdventureMessageFormatterFactory {
     final ImplicitConversion implicitConversion = ImplicitConversion.create();
     final PlaceholderContext placeholderContext = PlaceholderContext.create();
     final PlaceholderResolver placeholderResolver = PlaceholderResolverFactory.create();
-    final PlaceholderSanitizer placeholderSanitizer = DelegatingPlaceholderSanitizerFactory.create();
+    final PlaceholderSanitizer placeholderSanitizer = AdventurePlaceholderSanitizerFactory.create();
     final PlaceholderEvaluator placeholderEvaluator =
         ReflectivePlaceholderEvaluatorFactory.create();
     final PlaceholderProcessor placeholderProcessor =
