+++ honey-common/src/dev/shiza/honey/adventure/placeholder/sanitizer/AdventurePlaceholderSanitizerFactory.java
@@ -2,17 +2,15 @@ package dev.shiza.honey.adventure.placeholder.sanitizer;
 
 import dev.shiza.honey.placeholder.sanitizer.PlaceholderSanitizer;
 
-/** Factory for creating Adventure placeholder sanitizers. */
 public final class AdventurePlaceholderSanitizerFactory {
 
   private AdventurePlaceholderSanitizerFactory() {}
 
-  /**
-   * Creates a new instance of an Adventure placeholder sanitizer.
-   *
-   * @return a new instance of an Adventure placeholder sanitizer
-   */
   public static PlaceholderSanitizer create() {
     return new AdventurePlaceholderSanitizer();
+  }
+
+  public static PlaceholderSanitizer createReflective() {
+    return new AdventureReflectivePlaceholderSanitizer();
   }
 }
