+++ honey-common/test/dev/shiza/honey/placeholder/resolver/RegexPlaceholderResolverTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
-class RegexPlaceholderResolverTest {
+final class RegexPlaceholderResolverTest {
 
   @Test
   void resolveWithoutPlaceholders() {
